requirements.txt
agate==1.9.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
daff==1.4.2
dbt-adapters==1.16.0
dbt-common==1.25.1
dbt-core==1.10.3
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-protos==1.0.335
dbt-semantic-interfaces==0.8.5
deepdiff==7.0.1
distro==1.9.0
durationpy==0.10
fastapi==0.115.14
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
isodate==0.6.1
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
leather==0.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
mmh3==5.1.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numpy==2.3.1
oauthlib==3.3.1
onnxruntime==1.22.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
packaging==25.0
parsedatetime==2.6
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
posthog==5.4.0
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sqlparse==0.5.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
tqdm==4.67.1
transformers==4.53.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
zipp==3.23.0
# LangChain core packages (for LLM and agentic workflows)
langchain==0.3.26
langchain-community==0.3.27
langchain-openai==0.3.27
